fix: 前缀和计数改 uint16 修复无调整年溢出回绕
无调整年（空 off/work、无节日）为校验允许的合法配置，闰年且元旦为
周一~周四时普通工作日达 262 天，超出 uint8 上界 255：前缀和累计到
256 即回绕，全年统计 ordinary 得 6（262 mod 256），五键之和不再等于
total_days，MECE 不变量被破坏。原「至多 248 天」的上界估计漏算了
无调整年。

comboCounts 元素改 uint16 存储（每年内存 1.8KB→3.7KB，可忽略），
黑盒回归以 2024 空调整年固化 ordinary=262、weekend=104、五键之和
== Total；spec 前缀和 Requirement 与决策依据、docs/ARCHITECTURE.md
同步修订。

+++ calendar.go
@@ -50,10 +50,12 @@ func comboIndex(t DayType) int {
 }
 
 // comboCounts 5 种合法细粒度值各自的年内累计天数（按值计数，五值 MECE）。
-// 粗粒度计数可由其线性组合导出。单一类型年内天数有界（普通工作日至多
-// 248 天，其余类型更少），uint8 足以存储；跨年/多段累加不得直接在本类型
-// 上进行（会溢出），须先转入 comboTotals。
-type comboCounts [len(comboValues)]uint8
+// 粗粒度计数可由其线性组合导出。单一类型年内天数有界：无调整年（空 off/
+// work、无节日，校验允许）的普通工作日至多 262 天（闰年且元旦为周一~周四，
+// 366−104 周末），超出 uint8 上界 255——uint8 存储会在累计到 256 时回绕出
+// 错误统计，故用 uint16；跨年/多段累加仍不得直接在本类型上进行，须先转入
+// comboTotals。
+type comboCounts [len(comboValues)]uint16
 
 // diff 返回两组年内前缀计数的差（b-a，前缀和单调不减于被减数，无下溢）。
 func diffPrefix(b, a comboCounts) comboCounts {
@@ -64,8 +66,8 @@ func diffPrefix(b, a comboCounts) comboCounts {
 	return out
 }
 
-// comboTotals 跨年统计的宽类型累加器：各段年内差分（comboCounts，uint8）
-// 显式转入 int 后再相加，避免 uint8 多段直接累加溢出。
+// comboTotals 跨年统计的宽类型累加器：各段年内差分（comboCounts，uint16）
+// 显式转入 int 后再相加，避免 uint16 多段直接累加溢出。
 type comboTotals [len(comboValues)]int
 
 // add 累加一段年内差分计数（逐元素转 int）。
